Add files via upload
Tic tac toe is a game that uses a computer as the lead player and employs inside-out computing technology so that the computer may interact with the player as if they were in a human-to-human relationship. Tic tac toe is a two-player, zero-sum game, hence the optimal algorithm is an adversarial search algorithm, of which the negamax algorithm is one of the best examples. Negamax Algorithm is a variant of the algorithm In 2 minimax, max (a, b) = -min (-a, -b), meaning that what is good for X is bad for O, and vice versa.